Add files via upload
The Real estate app is a web-based real estate app in which the user can buy or rent the houses on the basis of various filters like purpose, Rent Frequency, Min and max 
price, location and so on an amazing Real Estate Application. With rental homes and for-sale properties, advanced property filtering, property details page with an image carousel.